+++ src/stack_mitosis/cli.clj
@@ -4,6 +4,7 @@
             [clojure.tools.cli :as cli]
             [clojure.tools.logging :as log]
             [stack-mitosis.interpreter :as interpreter]
+            [stack-mitosis.lookup :as lookup]
             [stack-mitosis.planner :as plan]
             [stack-mitosis.policy :as policy]
             [stack-mitosis.request :as r]
@@ -20,6 +21,7 @@
    ["-c" "--credentials FILENAME" "Credentials file in edn for iam assume-role"]
    ["-p" "--plan" "Display expected flightplan for operation."]
    ["-i" "--iam-policy" "Generate IAM policy for planned actions."]
+   [nil "--restore-snapshot" "Always clone using snapshot restore."]
    ["-h" "--help"]])
 
 (defn parse-args [args]
@@ -52,26 +54,36 @@
   (let [rds (interpreter/client)
         instances (interpreter/databases rds)]
     (when (interpreter/verify-databases-exist instances [source target])
-      (let [tags (interpreter/list-tags rds instances target)
-            plan (plan/replace-tree instances source target
-                                    :restart restart :tags tags)]
-        (cond (:plan options)
-              (do (println (flight-plan (interpreter/check-plan instances plan)))
-                  true)
-              (:iam-policy options)
-              (do (json/pprint (policy/from-plan instances plan))
-                  true)
-              :else
-              (let [last-action (interpreter/evaluate-plan rds plan)]
-                (not (contains? last-action :ErrorResponse))))))))
+      (let [same-vpc (lookup/same-vpc?
+                      (lookup/by-id instances source)
+                      (lookup/by-id instances target))
+            use-restore-snapshot (or (:restore-snapshot options) (not same-vpc))
+            source-snapshot (if use-restore-snapshot
+                              (interpreter/latest-snapshot rds source)
+                              nil)]
+
+        (when (or (not use-restore-snapshot)
+                  (interpreter/verify-snapshot-exists instances [source target]
+                                                      source-snapshot))
+          (let [tags (interpreter/list-tags rds instances target)
+                plan (plan/replace-tree instances source source-snapshot target
+                                        :restart restart :tags tags)]
+            (cond (:plan options)
+                  (do (println (flight-plan (interpreter/check-plan instances plan)))
+                      true)
+                  (:iam-policy options)
+                  (do (json/pprint (policy/from-plan instances plan))
+                      true)
+                  :else
+                  (let [last-action (interpreter/evaluate-plan rds plan)]
+                    (not (contains? last-action :ErrorResponse))))))))))
 
 (defn -main [& args]
   (let [{:keys [ok exit-msg] :as options} (parse-args args)]
     (when exit-msg
       (println exit-msg)
       (System/exit (if ok 0 1)))
-    (System/exit (if (process options) 0 1))
-    ))
+    (System/exit (if (process options) 0 1))))
 
 (comment
   (process (parse-args ["--source" "mitosis-prod" "--target" "mitosis-demo"