+++ src/stack_mitosis/operations.clj
@@ -38,6 +38,20 @@
                     {:SourceDBInstanceIdentifier source
                      :DBInstanceIdentifier replica})}))
 
+(defn restore-snapshot
+  ([snapshot-id source target] (restore-snapshot snapshot-id source target {}))
+  ([snapshot-id source target attributes]
+   {:op :RestoreDBInstanceFromDBSnapshot
+    :request (merge attributes
+                    {:DBSnapshotIdentifier snapshot-id
+                     :DBInstanceIdentifier target})
+    :meta {:SourceDBInstance source}}))
+
+(defn list-snapshots
+  ([target]
+   {:op :DescribeDBSnapshots
+    :request {:DBInstanceIdentifier target}}))
+
 (defn promote
   [id]
   {:op :PromoteReadReplica
@@ -68,7 +82,7 @@
 (defn blocking-operation?
   [action]
   (contains? #{:CreateDBInstance :CreateDBInstanceReadReplica
-               :PromoteReadReplica :ModifyDBInstance} (:op action)))
+               :PromoteReadReplica :ModifyDBInstance :RestoreDBInstanceFromDBSnapshot} (:op action)))
 
 (defn transition-to
   "Maps current rds status to in-progress, failed or done